added populate_initial function
